services/flask-backend/requirements.txt
# KillKrill Flask Backend Dependencies
# Python 3.13 compatible production-ready dependencies

# Web Framework
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Security-Too>=5.4.0
Flask-JWT-Extended>=4.5.0
Werkzeug>=3.0.0

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
PyMySQL>=1.1.0

# Redis Cache
redis>=5.2.1
hiredis>=2.3.2

# Authentication & Security
PyJWT>=2.10.1
passlib>=1.7.4
bcrypt>=4.2.1
cryptography>=44.0.1

# Data Validation & Serialization
pydantic>=2.10.0
python-multipart>=0.0.6
email-validator>=2.1.0

# HTTP & Networking
httpx>=0.27.0
aiohttp>=3.9.0
requests>=2.32.4

# gRPC
grpcio>=1.60.0
grpcio-tools>=1.60.0

# Monitoring & Observability
prometheus-client>=0.21.0
structlog>=24.0.0

# Configuration Management
python-decouple>=3.8

# Utilities
python-dateutil>=2.9.0
pytz>=2024.2
uuid6>=2024.1.12

# License Server Integration
requests>=2.32.4

# Production Server
gunicorn>=22.0.0
hypercorn[h3]>=0.16.0

# Development Dependencies (optional)
pytest>=8.3.4
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
black>=24.10.0
isort>=5.13.2
flake8>=7.1.1
mypy>=1.14.0
bandit>=1.7.5
netaddr>=0.10.0
pydal>=20241031.1
